+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Fixed
-- Fixed a PHP error that occured when programmatically changing an order's history in a queue job.
+- Fixed a PHP error that occurred when programmatically changing an order's history in a queue job.
 
 ## 3.2.7 - 2020-09-24