+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed a bug where the store location address had it‘s `isStoreLocation` property set to `false`. ([#1773](https://github.com/craftcms/commerce/issues/1773))
 - Fixed a bug where template suggestions weren’t showing on the product type template field.
 - Fixed a bug where deleting products would not delete related disabled variant. ([#1772](https://github.com/craftcms/commerce/issues/1772))
+- Fixed a bug where the incorrect confirmation message was shown when deleting a subscription plan.
 
 ## 3.2.7 - 2020-09-24