+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fixed a bug where the incorrect confirmation message was shown when deleting a subscription plan.
 - Fixed a PHP error that occurred when a subscription plan and gateway had been deleted. ([#1667](https://github.com/craftcms/commerce/issues/1667))
 - Fixed a bug where address edits weren’t persisting on the Edit Order page. ([#1766](https://github.com/craftcms/commerce/issues/1766))
+- Fixed a bug where coupon codes where were only applying if they were entered in a case-sensitive manor. ([#1763](https://github.com/craftcms/commerce/issues/1763))  
 
 ## 3.2.7 - 2020-09-24