+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added a “Variants” column to the Products index page. ([#1765](https://github.com/craftcms/commerce/issues/1765))
 - Added `craft\commerce\services\PaymentSources::getAllPaymentSourcesByGatewayId()`.
+- Added `craft\commerce\helpers\Purchasable`.
 
 ### Changed
 - Coupon codes are no longer case-sensitive. ([#1763](https://github.com/craftcms/commerce/issues/1763))
@@ -19,6 +20,7 @@
 - Fixed a bug where the incorrect confirmation message was shown when deleting a subscription plan.
 - Fixed a PHP error that occurred when a subscription plan and gateway had been deleted. ([#1667](https://github.com/craftcms/commerce/issues/1667))
 - Fixed a bug where address changes weren’t persisting on the Edit Order page. ([#1766](https://github.com/craftcms/commerce/issues/1766))  
+- Fixed a SQL error when saving a disabled product without populating the variant price, SKU, or stock required fields. ([#1764](https://github.com/craftcms/commerce/issues/1764))
 
 ## 3.2.7 - 2020-09-24