+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 ### Added
-- Added the `commerce/reset-data` console command. ([#581](https://github.com/craftcms/commerce/issues/581))
 - Added a “Variants” column to the Products index page. ([#1765](https://github.com/craftcms/commerce/issues/1765))
+- Added the `commerce/reset-data` command. ([#581](https://github.com/craftcms/commerce/issues/581))
 - Added `craft\commerce\console\controlllers\ResetData`.
 - Added `craft\commerce\elements\Variants::getSkuAsText()`.
 - Added `craft\commerce\helpers\Purchasable`.
@@ -23,9 +23,9 @@
 - Fixed a bug where the incorrect confirmation message was shown when deleting a subscription plan.
 - Fixed a PHP error that occurred when a subscription plan and gateway had been deleted. ([#1667](https://github.com/craftcms/commerce/issues/1667))
 - Fixed a bug where address changes weren’t persisting on the Edit Order page. ([#1766](https://github.com/craftcms/commerce/issues/1766))  
-- Fixed a SQL error when saving a disabled product without populating the variant price, SKU, or stock required fields. ([#1764](https://github.com/craftcms/commerce/issues/1764))
+- Fixed a SQL error that could occur when saving a product, if it was disabled and missing required field values. ([#1764](https://github.com/craftcms/commerce/issues/1764))
 - Fixed a bug where it was possible to change the primary currency after completed orders had been placed. ([#1777](https://github.com/craftcms/commerce/issues/1777))
-- Fixed a JavaScript error that occurred on the example template payment page.
+- Fixed a JavaScript error that occurred on the payment page of the example templates.
 
 ## 3.2.7 - 2020-09-24